+++ Store Sim 2/Player.h
@@ -9,7 +9,7 @@ public:
 	Player();
 
 	int getInventorySize() const { return _inventory.size(); }
-	float getMoney() const { return _money; }
+	double getMoney() const { return _money; }
 	Item *getItem(int itemPlace) { return _inventory[itemPlace - 1]; }
 	int getItemQuantity(int itemPlace) const { return _inventory[itemPlace - 1]->getItemQuantity(); }
 
@@ -18,11 +18,11 @@ public:
 
 	bool buyItem(Item *item, int quantity);
 
-	void removeMoney(float m){ _money -= floor(m * 100.00 + 0.5) / 100.00; }
-	void addMoney(float m) { _money += floor(m * 100.00 + 0.5) / 100.00; }
+	void removeMoney(double m){ _money -= floor(m * 100.00 + 0.5) / 100.00; }
+	void addMoney(double m) { _money += floor(m * 100.00 + 0.5) / 100.00; }
 
 private:
-	float _money = 100;
+	double _money = 100;
 	std::vector<Item*> _inventory;
